remove unneeded legacy client 'cushioning' code, wasnt required
