+++ src/validation.h
@@ -819,5 +819,6 @@ inline bool IsBlockPruned(const CBlockIndex* pblockindex)
 bool GetUTXOCoin(const COutPoint& outpoint, Coin& coin);
 int GetUTXOHeight(const COutPoint& outpoint);
 int GetUTXOConfirmations(const COutPoint& outpoint);
+bool IsLegacyMode();
 
 #endif // BITCOIN_VALIDATION_H
